+++ src/main/java/com/sim/wifi/upgrade/modules/permissions/service/AbUserCacheService.java
@@ -26,7 +26,7 @@ public interface AbUserCacheService {
     /**
      * 当角色相关资源信息改变时删除相关后台用户缓存
      */
-    void delResourceListByRole(Long roleId);
+    void delResourceListByRole(Integer roleId);
 
     /**
      * 当角色相关资源信息改变时删除相关后台用户缓存
